replace api_util and mappers to service layer
