+++ lib/core/connection_listener/connection_listener_cubit.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:shmr_finance/data/local/abstract/local_service.dart';
+import 'package:shmr_finance/data/local/abstract/local_repository.dart';
 import 'package:shmr_finance/core/logger/logger.dart';
 
 part 'connection_listener_state.dart';
