+++ lib/features/account/data/dto/create_account_use_case_request.dart
@@ -1,4 +1,3 @@
-import 'package:shmr_finance/data/dto/requests/export.dart';
 import 'package:shmr_finance/model/common_enums/currency_enum.dart';
 
 class CreateAccountUseCaseRequest {
@@ -10,7 +9,3 @@ class CreateAccountUseCaseRequest {
       {required this.name, required this.balance, required this.currency});
 }
 
-extension CreateAccountUseCaseRequestX on CreateAccountUseCaseRequest {
-  ApiAccountCreateRequest toData() => ApiAccountCreateRequest(
-      name: name, balance: balance.toString(), currency: currency.code);
-}
