+++ lib/service/mappers/api/api_account_mapper.dart
@@ -1,7 +1,7 @@
 import 'package:shmr_finance/features/account/domain/entities/account.dart';
 import 'package:shmr_finance/model/common_enums/currency_enum.dart';
 
-import '../dto/responses/account_dto/account_dto.dart';
+import '../../../data/dto/responses/account_dto/account_dto.dart';
 
 extension ApiAccountX on ApiAccount {
   Account toDomain() => Account(
